+++ frontend/lib/api/types/response.ts
@@ -47,91 +47,6 @@ export interface RecipeSearchQuery {
   requireAllFoods?: boolean;
   search?: string | null;
 }
-export interface RecipeSuggestionQuery {
-  orderBy?: string | null;
-  orderByNullPosition?: OrderByNullPosition | null;
-  orderDirection?: OrderDirection;
-  queryFilter?: string | null;
-  paginationSeed?: string | null;
-  limit?: number;
-  maxMissingFoods?: number;
-  maxMissingTools?: number;
-  includeFoodsOnHand?: boolean;
-  includeToolsOnHand?: boolean;
-}
-export interface RecipeSuggestionResponse {
-  items: RecipeSuggestionResponseItem[];
-}
-export interface RecipeSuggestionResponseItem {
-  recipe: RecipeSummary;
-  missingFoods: IngredientFood[];
-  missingTools: RecipeTool[];
-}
-export interface RecipeSummary {
-  id?: string | null;
-  userId?: string;
-  householdId?: string;
-  groupId?: string;
-  name?: string | null;
-  slug?: string;
-  image?: unknown;
-  recipeYield?: string | null;
-  totalTime?: string | null;
-  prepTime?: string | null;
-  cookTime?: string | null;
-  performTime?: string | null;
-  description?: string | null;
-  recipeCategory?: RecipeCategory[] | null;
-  tags?: RecipeTag[] | null;
-  tools?: RecipeTool[];
-  rating?: number | null;
-  orgURL?: string | null;
-  dateAdded?: string | null;
-  dateUpdated?: string | null;
-  createdAt?: string | null;
-  updatedAt?: string | null;
-  lastMade?: string | null;
-}
-export interface RecipeCategory {
-  id?: string | null;
-  name: string;
-  slug: string;
-}
-export interface RecipeTag {
-  id?: string | null;
-  name: string;
-  slug: string;
-}
-export interface RecipeTool {
-  id: string;
-  name: string;
-  slug: string;
-  onHand?: boolean;
-}
-export interface IngredientFood {
-  id: string;
-  name: string;
-  pluralName?: string | null;
-  description?: string;
-  extras?: {
-    [k: string]: unknown;
-  } | null;
-  onHand?: boolean;
-  labelId?: string | null;
-  aliases?: IngredientFoodAlias[];
-  label?: MultiPurposeLabelSummary | null;
-  createdAt?: string | null;
-  updatedAt?: string | null;
-}
-export interface IngredientFoodAlias {
-  name: string;
-}
-export interface MultiPurposeLabelSummary {
-  name: string;
-  color?: string;
-  groupId: string;
-  id: string;
-}
 export interface RequestQuery {
   orderBy?: string | null;
   orderByNullPosition?: OrderByNullPosition | null;
