+++ mealie/schema/recipe/__init__.py
@@ -75,6 +75,7 @@ from .recipe_scraper import ScrapeRecipe, ScrapeRecipeBase, ScrapeRecipeData, Sc
 from .recipe_settings import RecipeSettings
 from .recipe_share_token import RecipeShareToken, RecipeShareTokenCreate, RecipeShareTokenSave, RecipeShareTokenSummary
 from .recipe_step import IngredientReferences, RecipeStep
+from .recipe_suggestion import RecipeSuggestionQuery, RecipeSuggestionResponse, RecipeSuggestionResponseItem
 from .recipe_timeline_events import (
     RecipeTimelineEventCreate,
     RecipeTimelineEventIn,
@@ -109,6 +110,9 @@ __all__ = [
     "RecipeTimelineEventUpdate",
     "TimelineEventImage",
     "TimelineEventType",
+    "RecipeSuggestionQuery",
+    "RecipeSuggestionResponse",
+    "RecipeSuggestionResponseItem",
     "Nutrition",
     "RecipeShareToken",
     "RecipeShareTokenCreate",
