+++ mealie/schema/response/__init__.py
@@ -17,7 +17,6 @@ from .query_filter import (
     RelationalOperator,
 )
 from .query_search import SearchFilter
-from .recipe_filter import RecipeSuggestionQuery, RecipeSuggestionResponse, RecipeSuggestionResponseItem
 from .responses import ErrorResponse, FileTokenResponse, SuccessResponse
 from .validation import ValidationResponse
 
@@ -40,7 +39,4 @@ __all__ = [
     "ErrorResponse",
     "FileTokenResponse",
     "SuccessResponse",
-    "RecipeSuggestionQuery",
-    "RecipeSuggestionResponse",
-    "RecipeSuggestionResponseItem",
 ]
